feat(e2e): finish support layout and delete bulk-import-shared
